make the `pub` command have pass-through arguments
PubCommand now uses a custom ArgParser so that it may work seamlessly
with Command and CommandRunner even though those classes try to work
with a 'help' flag on PubCommand. The custom ArgParser is based on the
`allowAnything()` ArgParser to facilitate pass-through, but it ignores
adding the 'help' flag and pretends it exists when needed to match the
behavior CommandRunner expects.