catch ProcessExceptions thrown by the `pub` command
If Flutter isn't installed but the `pub` command thinks it needs it, an
unhandled ProcessException makes its way up and crashes everything.
There's some protection because it would be hard to get mono_repo
installed without `pub`, but this issue comes up with `flutter` and
testing: `flutter` isn't a devDependency, but the new tests test that
it can be used when needed. Those tests don't require `flutter` to
work, but they do need `mono_repo` to not halt and catch fire if it
doesn't.