Parse CIJob maps once
Move the `parse` operation out of the loop across OS and SDK
combinations. Change from a factory to a static method that returns a
record with the two fields that get parsed out of the yaml map at this
level. The `parse` method creates a copy of the map, which shouldn't be
necessary to do more than once.